Moving more functions to utils. basic_dqn is updated to match with these new changes.
